Put ghrel after AGR and AUR due to grep url basename limitation
this handles packages from a list of urls only which can not accurately determine a package name, to avoid false positive or negative results, this has to be done at the last order before falling to package not found.

Since it is possible to build a pacman repo easily using github release, this feature is not much needed.